#65 Move the multiplexing algebra to a MuxEndpoints trait
