+++ akka-http/client/src/main/scala/endpoints/akkahttp/client/Endpoints.scala
@@ -2,7 +2,6 @@ package endpoints.akkahttp.client
 
 import akka.http.scaladsl.model._
 import akka.stream.Materializer
-import endpoints.algebra.{Decoder, Encoder, MuxRequest}
 import endpoints.{Tupler, algebra}
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -59,35 +58,10 @@ class Endpoints(val settings: EndpointsSettings)
         result <- response(resp).flatMap(futureFromEither)
       } yield result
 
-  private def futureFromEither[A](errorOrA: Either[Throwable, A]): Future[A] =
+  private[client] def futureFromEither[A](errorOrA: Either[Throwable, A]): Future[A] =
     errorOrA match {
       case Left(error) => Future.failed(error)
       case Right(a) => Future.successful(a)
     }
-
-  class MuxEndpoint[Req <: algebra.MuxRequest, Resp, Transport](
-                                                                 request: Request[Transport],
-                                                                 response: Response[Transport]
-                                                               ) {
-    def apply(
-               req: Req
-             )(implicit
-               encoder: Encoder[Req, Transport],
-               decoder: Decoder[Transport, Resp]
-             ): Future[req.Response] =
-      request(encoder.encode(req)).flatMap { resp =>
-        response(resp).flatMap { t =>
-          futureFromEither(t).flatMap(tt =>
-            futureFromEither(decoder.decode(tt)).map(_.asInstanceOf[req.Response])
-          )
-        }
-      }
-  }
-
-  def muxEndpoint[Req <: MuxRequest, Resp, Transport](
-                                                       request: Request[Transport],
-                                                       response: Response[Transport]
-                                                     ): MuxEndpoint[Req, Resp, Transport] =
-    new MuxEndpoint[Req, Resp, Transport](request, response)
 
 }
