+++ akka-http/server/src/main/scala/endpoints/akkahttp/server/Endpoints.scala
@@ -2,7 +2,6 @@ package endpoints.akkahttp.server
 
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.{Directive1, Directives, Route}
-import endpoints.algebra.{Decoder, Encoder, MuxRequest}
 import endpoints.{Tupler, algebra}
 
 import scala.concurrent.Future
@@ -65,59 +64,4 @@ trait Endpoints extends algebra.Endpoints with Urls with Methods {
 
   def endpoint[A, B](request: Request[A], response: Response[B]): Endpoint[A, B] = Endpoint(request, response)
 
-
-  class MuxEndpoint[Req <: MuxRequest, Resp, Transport](request: Request[Transport], response: Response[Transport]) {
-
-    def implementedBy(handler: MuxHandler[Req, Resp])(implicit
-      decoder: Decoder[Transport, Req],
-      encoder: Encoder[Resp, Transport]
-    ): Route = handleAsync(req => Future.successful(handler(req)))
-
-    def implementedByAsync(handler: MuxHandlerAsync[Req, Resp])(implicit
-      decoder: Decoder[Transport, Req],
-      encoder: Encoder[Resp, Transport]
-    ): Route = handleAsync(req => handler(req))
-
-    private def handleAsync(handler: Req {type Response = Resp} => Future[Resp])(implicit
-      decoder: Decoder[Transport, Req],
-      encoder: Encoder[Resp, Transport]
-    ): Route =
-      request { request =>
-        Directives.onComplete(handler(decoder.decode(request).right.get /* TODO Handle failure */ .asInstanceOf[Req {type Response = Resp}])) {
-          case Success(result) => response(encoder.encode(result))
-          case Failure(ex) => Directives.complete(ex)
-        }
-      }
-
-  }
-
-  def muxEndpoint[Req <: MuxRequest, Resp, Transport](
-    request: Request[Transport],
-    response: Response[Transport]
-  ): MuxEndpoint[Req, Resp, Transport] =
-    new MuxEndpoint[Req, Resp, Transport](request, response)
-
-
-}
-
-/**
-  * A function whose return type depends on the type
-  * of the given `req`.
-  *
-  * @tparam Req  Request base type
-  * @tparam Resp Response base type
-  */
-trait MuxHandlerAsync[Req <: MuxRequest, Resp] {
-  def apply[R <: Resp](req: Req {type Response = R}): Future[R]
-}
-
-/**
-  * A function whose return type depends on the type
-  * of the given `req`.
-  *
-  * @tparam Req  Request base type
-  * @tparam Resp Response base type
-  */
-trait MuxHandler[Req <: MuxRequest, Resp] {
-  def apply[R <: Resp](req: Req {type Response = R}): R
 }
