+++ documentation/examples/cqrs/public-server/src/main/scala/cqrs/publicserver/QueriesClient.scala
@@ -1,6 +1,6 @@
 package cqrs.publicserver
 
-import endpoints.play.client.{CirceEntities, Endpoints}
+import endpoints.play.client.{CirceEntities, Endpoints, MuxEndpoints}
 import cqrs.queries.QueriesEndpoints
 import play.api.libs.ws.WSClient
 
@@ -9,4 +9,5 @@ import scala.concurrent.ExecutionContext
 class QueriesClient(baseUrl: String, wsClient: WSClient)(implicit ec: ExecutionContext)
   extends Endpoints(baseUrl, wsClient)
     with CirceEntities
+    with MuxEndpoints
     with QueriesEndpoints
