+++ openapi/openapi/src/main/scala/endpoints/documented/algebra/Endpoints.scala
@@ -2,8 +2,6 @@ package endpoints
 package documented
 package algebra
 
-import endpoints.algebra.MuxRequest
-
 import scala.language.higherKinds
 
 /**
@@ -31,29 +29,5 @@ trait Endpoints
     * @param response Response
     */
   def endpoint[A, B](request: Request[A], response: Response[B]): Endpoint[A, B]
-
-  /**
-    * Information carried by a multiplexed HTTP endpoint.
-    */
-  type MuxEndpoint[Req <: MuxRequest, Resp, Transport]
-
-  /**
-    * Multiplexed HTTP endpoint.
-    *
-    * A multiplexing endpoint makes it possible to use several request
-    * and response types in the same HTTP endpoint. In other words, it
-    * allows to define several different actions through a singe HTTP
-    * endpoint.
-    *
-    * @param request The request
-    * @param response The response
-    * @tparam Req The base type of possible requests
-    * @tparam Resp The base type of possible responses
-    * @tparam Transport The data type used to transport the requests and responses
-    */
-  def muxEndpoint[Req <: MuxRequest, Resp, Transport](
-    request: Request[Transport],
-    response: Response[Transport]
-  ): MuxEndpoint[Req, Resp, Transport]
 
 }
