+++ play/server/src/main/scala/endpoints/play/server/Endpoints.scala
@@ -2,7 +2,6 @@ package endpoints.play.server
 
 import endpoints.algebra
 import endpoints.Tupler
-import endpoints.algebra.{Decoder, Encoder, MuxRequest}
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import play.api.libs.functional.InvariantFunctor
 import play.api.libs.functional.syntax._
@@ -231,47 +230,6 @@ trait Endpoints extends algebra.Endpoints with Urls with Methods {
   def endpoint[A, B](request: Request[A], response: Response[B]): Endpoint[A, B] =
     Endpoint(request, response)
 
-  class MuxEndpoint[Req <: MuxRequest, Resp, Transport](
-    request: Request[Transport],
-    response: Response[Transport]
-  ) {
-    def implementedBy(
-      handler: MuxHandler[Req, Resp]
-    )(implicit
-      decoder: Decoder[Transport, Req],
-      encoder: Encoder[Resp, Transport]
-    ): ToPlayHandler =
-      toPlayHandler(req => Future.successful(handler(req)))
-
-    def implementedByAsync(
-      handler: MuxHandlerAsync[Req, Resp]
-    )(implicit
-      decoder: Decoder[Transport, Req],
-      encoder: Encoder[Resp, Transport]
-    ): ToPlayHandler =
-      toPlayHandler(req => handler(req))
-
-    def toPlayHandler(
-      handler: Req { type Response = Resp } => Future[Resp]
-    )(implicit
-      decoder: Decoder[Transport, Req],
-      encoder: Encoder[Resp, Transport]
-    ): ToPlayHandler =
-      header =>
-        request.decode(header).map { bodyParser =>
-          Action.async(bodyParser) { request =>
-            handler(decoder.decode(request.body).right.get /* TODO Handle failure */.asInstanceOf[Req { type Response = Resp}])
-              .map(resp => response(encoder.encode(resp)))
-          }
-        }
-  }
-
-  def muxEndpoint[Req <: MuxRequest, Resp, Transport](
-    request: Request[Transport],
-    response: Transport => Result
-  ): MuxEndpoint[Req, Resp, Transport] =
-    new MuxEndpoint[Req, Resp, Transport](request, response)
-
   /**
     * Builds a Play router out of endpoint definitions.
     *
@@ -294,28 +252,4 @@ trait Endpoints extends algebra.Endpoints with Urls with Methods {
   implicit def EmptyEndpointToPlayHandler[A, B](endpoint: Endpoint[A, B])(implicit ev: Unit =:= B): ToPlayHandler =
     endpoint.implementedBy(_ => ())
 
-}
-
-//#mux-handler-async
-/**
-  * A function whose return type depends on the type
-  * of the given `req`.
-  *
-  * @tparam Req Request base type
-  * @tparam Resp Response base type
-  */
-trait MuxHandlerAsync[Req <: MuxRequest, Resp] {
-  def apply[R <: Resp](req: Req { type Response = R }): Future[R]
-}
-//#mux-handler-async
-
-/**
-  * A function whose return type depends on the type
-  * of the given `req`.
-  *
-  * @tparam Req Request base type
-  * @tparam Resp Response base type
-  */
-trait MuxHandler[Req <: MuxRequest, Resp] {
-  def apply[R <: Resp](req: Req { type Response = R }): R
 }
