+++ xhr/client/src/main/scala/endpoints/xhr/thenable/Endpoints.scala
@@ -1,6 +1,5 @@
 package endpoints.xhr.thenable
 
-import endpoints.algebra.{Decoder, Encoder, MuxRequest}
 import endpoints.xhr
 
 import scala.scalajs.js
@@ -21,30 +20,5 @@ trait Endpoints extends xhr.Endpoints {
           xhr => error(xhr.responseText)
         )
       })
-
-  class MuxEndpoint[Req <: MuxRequest, Resp, Transport](
-    request: Request[Transport],
-    response: Response[Transport]
-  ) {
-    def apply(
-      req: Req
-    )(implicit
-      encoder: Encoder[Req, Transport],
-      decoder: Decoder[Transport, Resp]
-    ): js.Thenable[req.Response] = {
-      new js.Promise[req.Response]((resolve, error) => {
-        muxPerformXhr(request, response, req)(
-          _.fold(exn => error(exn.getMessage), resp => resolve(resp)),
-          xhr => error(xhr.responseText)
-        )
-      })
-    }
-  }
-
-  def muxEndpoint[Req <: MuxRequest, Resp, Transport](
-    request: Request[Transport],
-    response: Response[Transport]
-  ): MuxEndpoint[Req, Resp, Transport] =
-    new MuxEndpoint[Req, Resp, Transport](request, response)
 
 }
